// src/app/interfaces/iempleado.ts
export interface IEmpleado {
    id: number
    nombre: string;
    apellidos: string;
    email: string;
    telefono: string;
    fechaNacimiento: Date;
    fechaContratacion: Date;
    salario: number;
    numeroSS: string;
    direccion: string;
    imagen: string;
    departamento: string;
    evaluacion: number;

}